[Security Solution] move components and hooks around and clean up codeowners
