+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_creation_ui/components/rule_preview/use_preview_rule.ts
@@ -14,7 +14,7 @@ import type {
 } from '../../../../../common/api/detection_engine';
 import { useKibana } from '../../../../common/lib/kibana';
 import { previewRule } from '../../../rule_management/api/api';
-import { transformOutput } from '../../../../detections/containers/detection_engine/rules/transforms';
+import { transformOutput } from '../../../common/transforms';
 import type { TimeframePreviewOptions } from '../../../common/types';
 import { usePreviewInvocationCount } from './use_preview_invocation_count';
 import * as i18n from './translations';
