+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_exceptions/logic/use_close_alerts.tsx
@@ -9,10 +9,10 @@ import { useEffect, useRef, useState } from 'react';
 import type { estypes } from '@elastic/elasticsearch';
 import type { ExceptionListItemSchema } from '@kbn/securitysolution-io-ts-list-types';
 import {
-  buildAlertStatusesFilter,
   buildAlertsFilterByRuleIds,
+  buildAlertStatusesFilter,
 } from '../../../detections/components/alerts_table/default_config';
-import { getEsQueryFilter } from '../../../detections/containers/detection_engine/exceptions/get_es_query_filter';
+import { getEsQueryFilter } from '../utils/get_es_query_filter';
 import type { IndexPatternArray } from '../../../../common/api/detection_engine/model/rule_schema';
 import { prepareExceptionItemsForBulkClose } from '../utils/helpers';
 import * as i18nCommon from '../../../common/translations';
