+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_management/api/hooks/use_create_rule_mutation.ts
@@ -4,6 +4,7 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
+
 import type { UseMutationOptions } from '@tanstack/react-query';
 import { useMutation } from '@tanstack/react-query';
 import { DETECTION_ENGINE_RULES_URL } from '../../../../../common/constants';
@@ -11,7 +12,7 @@ import type {
   RuleCreateProps,
   RuleResponse,
 } from '../../../../../common/api/detection_engine/model/rule_schema';
-import { transformOutput } from '../../../../detections/containers/detection_engine/rules/transforms';
+import { transformOutput } from '../../../common/transforms';
 import { createRule } from '../api';
 import { useInvalidateFetchRuleManagementFiltersQuery } from './use_fetch_rule_management_filters_query';
 import { useInvalidateFindRulesQuery } from './use_find_rules_query';
