+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_management/api/hooks/use_update_rule_mutation.ts
@@ -4,13 +4,14 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
+
 import type { UseMutationOptions } from '@tanstack/react-query';
 import { useMutation } from '@tanstack/react-query';
 import type {
   RuleResponse,
   RuleUpdateProps,
 } from '../../../../../common/api/detection_engine/model/rule_schema';
-import { transformOutput } from '../../../../detections/containers/detection_engine/rules/transforms';
+import { transformOutput } from '../../../common/transforms';
 import { DETECTION_ENGINE_RULES_URL } from '../../../../../common/constants';
 import { updateRule } from '../api';
 import { useInvalidateFindRulesQuery } from './use_find_rules_query';
