+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_management/components/rule_details/three_way_diff/final_edit/fields/kql_query/kql_query_edit.tsx
@@ -15,8 +15,8 @@ import { HiddenField, UseField } from '../../../../../../../../shared_imports';
 import { QueryBarField } from '../../../../../../../rule_creation_ui/components/query_bar_field';
 import * as stepDefineRuleI18n from '../../../../../../../rule_creation_ui/components/step_define_rule/translations';
 import type { DiffableRule } from '../../../../../../../../../common/api/detection_engine';
-import type { SetRuleQuery } from '../../../../../../../../detections/containers/detection_engine/rules/use_rule_from_timeline';
-import { useRuleFromTimeline } from '../../../../../../../../detections/containers/detection_engine/rules/use_rule_from_timeline';
+import type { SetRuleQuery } from '../../../../../../../../detections/hooks/use_rule_from_timeline';
+import { useRuleFromTimeline } from '../../../../../../../../detections/hooks/use_rule_from_timeline';
 import { useGetSavedQuery } from '../../../../../../../common/use_get_saved_query';
 import type { RuleFieldEditComponentProps } from '../../../field_final_side';
 import { useDiffableRuleDataView } from '../hooks/use_diffable_rule_data_view';
