+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_management/hooks/use_missing_privileges.test.tsx
@@ -8,11 +8,11 @@
 import { renderHook } from '@testing-library/react';
 import { useMissingPrivileges } from './use_missing_privileges';
 import { useUserPrivileges } from '../../../common/components/user_privileges';
-import { useUserData } from '../../components/user_info';
+import { useUserData } from '../../../detections/components/user_info';
 import { SECURITY_FEATURE_ID } from '../../../../common';
 
 jest.mock('../../../common/components/user_privileges');
-jest.mock('../../components/user_info');
+jest.mock('../../../detections/components/user_info');
 
 const detectionEnginePrivileges = {
   username: 'elastic',
