+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_management/hooks/use_missing_privileges.ts
@@ -7,8 +7,8 @@
 
 import { useMemo } from 'react';
 import { SECURITY_FEATURE_ID } from '../../../../common/constants';
-import type { Privilege } from '../../containers/detection_engine/alerts/types';
-import { useUserData } from '../../components/user_info';
+import type { Privilege } from '../../../detections/containers/detection_engine/alerts/types';
+import { useUserData } from '../../../detections/components/user_info';
 import { useUserPrivileges } from '../../../common/components/user_privileges';
 
 const REQUIRED_INDEX_PRIVILEGES = ['read', 'write', 'view_index_metadata', 'manage'] as const;
