+++ x-pack/solutions/security/plugins/security_solution/public/detections/hooks/use_rule_from_timeline.test.ts
@@ -6,32 +6,31 @@
  */
 
 import { act, renderHook, waitFor } from '@testing-library/react';
-
 import { useRuleFromTimeline } from './use_rule_from_timeline';
-import { useGetInitialUrlParamValue } from '../../../../common/utils/global_query_string/helpers';
-import { resolveTimeline } from '../../../../timelines/containers/api';
-import { useSourcererDataView } from '../../../../sourcerer/containers';
-import { mockSourcererScope } from '../../../../sourcerer/containers/mocks';
-import { useAppToasts } from '../../../../common/hooks/use_app_toasts';
-import { useAppToastsMock } from '../../../../common/hooks/use_app_toasts.mock';
-import { mockTimeline } from '../../../../../server/lib/timeline/__mocks__/create_timelines';
-import type { TimelineModel } from '../../../..';
-import type { ResolveTimelineResponse } from '../../../../../common/api/timeline';
-import { useDataView } from '../../../../data_view_manager/hooks/use_data_view';
+import { useGetInitialUrlParamValue } from '../../common/utils/global_query_string/helpers';
+import { resolveTimeline } from '../../timelines/containers/api';
+import { useSourcererDataView } from '../../sourcerer/containers';
+import { mockSourcererScope } from '../../sourcerer/containers/mocks';
+import { useAppToasts } from '../../common/hooks/use_app_toasts';
+import { useAppToastsMock } from '../../common/hooks/use_app_toasts.mock';
+import { mockTimeline } from '../../../server/lib/timeline/__mocks__/create_timelines';
+import type { TimelineModel } from '../..';
+import type { ResolveTimelineResponse } from '../../../common/api/timeline';
+import { useDataView } from '../../data_view_manager/hooks/use_data_view';
 import {
   getMockDataView,
   getMockDataViewWithMatchedIndices,
-} from '../../../../data_view_manager/mocks/mock_data_view';
-import { withIndices } from '../../../../data_view_manager/hooks/__mocks__/use_data_view';
+} from '../../data_view_manager/mocks/mock_data_view';
+import { withIndices } from '../../data_view_manager/hooks/__mocks__/use_data_view';
 
-jest.mock('../../../../common/hooks/use_experimental_features');
-jest.mock('../../../../common/utils/global_query_string/helpers');
-jest.mock('../../../../timelines/containers/api');
-jest.mock('../../../../common/hooks/use_app_toasts');
-jest.mock('../../../../sourcerer/containers');
-jest.mock('../../../../common/components/discover_in_timeline/use_discover_in_timeline_context');
-jest.mock('../../../../common/components/link_to', () => {
-  const originalModule = jest.requireActual('../../../../common/components/link_to');
+jest.mock('../../common/hooks/use_experimental_features');
+jest.mock('../../common/utils/global_query_string/helpers');
+jest.mock('../../timelines/containers/api');
+jest.mock('../../common/hooks/use_app_toasts');
+jest.mock('../../sourcerer/containers');
+jest.mock('../../common/components/discover_in_timeline/use_discover_in_timeline_context');
+jest.mock('../../common/components/link_to', () => {
+  const originalModule = jest.requireActual('../../common/components/link_to');
   return {
     ...originalModule,
     getTimelineUrl: jest.fn(),
@@ -40,6 +39,7 @@ jest.mock('../../../../common/components/link_to', () => {
     }),
   };
 });
+jest.mock('../../data_view_manager/hooks/use_data_view');
 
 const mockDispatch = jest.fn();
 jest.mock('react-redux', () => {
