+++ x-pack/solutions/security/plugins/security_solution/public/entity_analytics/components/risk_engine_privileges_callout/translations.tsx
@@ -11,7 +11,7 @@ import { FormattedMessage } from '@kbn/i18n-react';
 import React from 'react';
 import type { MissingPrivileges } from '../../../../common/entity_analytics/risk_engine';
 import { useKibana } from '../../../common/lib/kibana';
-import { CommaSeparatedValues } from '../../../detections/components/callouts/comma_separated_values';
+import { CommaSeparatedValues } from '../../../detection_engine/rule_management/components/callouts/comma_separated_values';
 
 export const MISSING_PRIVILEGES_CALLOUT_TITLE = i18n.translate(
   'xpack.securitySolution.riskEngine.missingPrivilegesCallOut.messageTitle',
