+++ x-pack/solutions/security/plugins/security_solution/public/siem_migrations/rules/pages/index.tsx
@@ -23,7 +23,7 @@ import { SecuritySolutionPageWrapper } from '../../../common/components/page_wra
 import { SecurityPageName } from '../../../app/types';
 
 import { MigrationRulesTable } from '../components/rules_table';
-import { NeedAdminForUpdateRulesCallOut } from '../../../detections/components/callouts/need_admin_for_update_rules_callout';
+import { NeedAdminForUpdateRulesCallOut } from '../../../detection_engine/rule_management/components/callouts/need_admin_for_update_rules_callout';
 import { MissingPrivilegesCallOut } from './missing_privileges_callout';
 import { HeaderButtons, UnknownMigration } from '../../common/components';
 import { useLatestStats } from '../service/hooks/use_latest_stats';
